+++ CHANGES.md
@@ -1,5 +1,14 @@
 # Changes History
 
+2.1.1
+------
+Change catching type in Repository constructor from \Exception to \Throwable
+Add description to Repository constructor
+Update accepted exception type in RepositoryException from \Exception to \Throwable.
+Add Mocker helper yo use db connections mocks instead sqlite:memory.
+Add mockery/mockery in composer dependencies.
+Add base repository test.
+
 2.1.0
 ------
 Add EloquentRepository with extended method **getWith()** to retrieve list of items